Revert "[Breaking] Cleanup and fixes (#387)"
This reverts commit e07a25f742d94877adbad4e90531ddf6fb1f044b.

+++ android/src/main/java/fr/greweb/reactnativeviewshot/RNViewShotModule.java
@@ -3,20 +3,23 @@ package fr.greweb.reactnativeviewshot;
 
 import android.app.Activity;
 import android.content.Context;
-import androidx.annotation.NonNull;
-
 import android.net.Uri;
+import android.os.AsyncTask;
+import androidx.annotation.NonNull;
 import android.util.DisplayMetrics;
 import android.util.Log;
 
+import com.facebook.react.bridge.GuardedAsyncTask;
 import com.facebook.react.bridge.Promise;
 import com.facebook.react.bridge.ReactApplicationContext;
+import com.facebook.react.bridge.ReactContext;
 import com.facebook.react.bridge.ReactContextBaseJavaModule;
 import com.facebook.react.bridge.ReactMethod;
 import com.facebook.react.bridge.ReadableMap;
 import com.facebook.react.uimanager.UIManagerModule;
 
 import java.io.File;
+import java.io.FilenameFilter;
 import java.io.IOException;
 import java.util.Collections;
 import java.util.Map;
@@ -35,7 +38,6 @@ public class RNViewShotModule extends ReactContextBaseJavaModule {
         this.reactContext = reactContext;
     }
 
-    @NonNull
     @Override
     public String getName() {
         return RNVIEW_SHOT;
@@ -46,13 +48,22 @@ public class RNViewShotModule extends ReactContextBaseJavaModule {
         return Collections.emptyMap();
     }
 
+    @Override
+    public void onCatalystInstanceDestroy() {
+        super.onCatalystInstanceDestroy();
+        new CleanTask(getReactApplicationContext()).executeOnExecutor(AsyncTask.THREAD_POOL_EXECUTOR);
+    }
+
     @ReactMethod
     public void releaseCapture(String uri) {
         final String path = Uri.parse(uri).getPath();
         if (path == null) return;
         File file = new File(path);
         if (!file.exists()) return;
-        file.delete();
+        File parent = file.getParentFile();
+        if (parent.equals(reactContext.getExternalCacheDir()) || parent.equals(reactContext.getCacheDir())) {
+            file.delete();
+        }
     }
 
     @ReactMethod
@@ -60,10 +71,7 @@ public class RNViewShotModule extends ReactContextBaseJavaModule {
         final ReactApplicationContext context = getReactApplicationContext();
         final DisplayMetrics dm = context.getResources().getDisplayMetrics();
 
-        String extension = options.getString("format");
-        if (extension == null) {
-            extension = "jpg";
-        }
+        final String extension = options.getString("format");
         final int imageFormat = "jpg".equals(extension)
                 ? Formats.JPEG
                 : "webm".equals(extension)
@@ -77,28 +85,21 @@ public class RNViewShotModule extends ReactContextBaseJavaModule {
         final Integer scaleHeight = options.hasKey("height") ? (int) (dm.density * options.getDouble("height")) : null;
         final String resultStreamFormat = options.getString("result");
         final Boolean snapshotContentContainer = options.getBoolean("snapshotContentContainer");
-        final String path = options.getString("path");
 
         try {
             File outputFile = null;
-            if (Results.FILE.equals(resultStreamFormat)) {
-                if (path == null || path.isEmpty()) {
-                    outputFile = createTempFile(getReactApplicationContext(), extension);
-                } else {
-                    outputFile = new File(path);
-                }
+            if (Results.TEMP_FILE.equals(resultStreamFormat)) {
+                outputFile = createTempFile(getReactApplicationContext(), extension);
             }
 
             final Activity activity = getCurrentActivity();
             final UIManagerModule uiManager = this.reactContext.getNativeModule(UIManagerModule.class);
 
-            if (uiManager != null) {
-                uiManager.addUIBlock(new ViewShot(
-                        tag, extension, imageFormat, quality,
-                        scaleWidth, scaleHeight, outputFile, resultStreamFormat,
-                        snapshotContentContainer, reactContext, activity, promise)
-                );
-            }
+            uiManager.addUIBlock(new ViewShot(
+                    tag, extension, imageFormat, quality,
+                    scaleWidth, scaleHeight, outputFile, resultStreamFormat,
+                    snapshotContentContainer, reactContext, activity, promise)
+            );
         } catch (final Throwable ex) {
             Log.e(RNVIEW_SHOT, "Failed to snapshot view tag " + tag, ex);
             promise.reject(ViewShot.ERROR_UNABLE_TO_SNAPSHOT, "Failed to snapshot view tag " + tag);
@@ -110,14 +111,75 @@ public class RNViewShotModule extends ReactContextBaseJavaModule {
         captureRef(-1, options, promise);
     }
 
-    private static final String TEMP_FILE_PREFIX = "react-native-view-shot";
+    private static final String TEMP_FILE_PREFIX = "ReactNative-snapshot-image";
 
     /**
-     * Create a temporary file in the cache directory.
+     * Asynchronous task that cleans up cache dirs (internal and, if available, external) of cropped
+     * image files. This is run when the catalyst instance is being destroyed (i.e. app is shutting
+     * down) and when the module is instantiated, to handle the case where the app crashed.
+     */
+    private static class CleanTask extends GuardedAsyncTask<Void, Void> implements FilenameFilter {
+        private final File cacheDir;
+        private final File externalCacheDir;
+
+        private CleanTask(ReactContext context) {
+            super(context);
+
+            cacheDir = context.getCacheDir();
+            externalCacheDir = context.getExternalCacheDir();
+        }
+
+        @Override
+        protected void doInBackgroundGuarded(Void... params) {
+            if (null != cacheDir) {
+                cleanDirectory(cacheDir);
+            }
+
+            if (externalCacheDir != null) {
+                cleanDirectory(externalCacheDir);
+            }
+        }
+
+        @Override
+        public final boolean accept(File dir, String filename) {
+            return filename.startsWith(TEMP_FILE_PREFIX);
+        }
+
+        private void cleanDirectory(@NonNull final File directory) {
+            final File[] toDelete = directory.listFiles(this);
+
+            if (toDelete != null) {
+                for (File file : toDelete) {
+                    if (file.delete()) {
+                        Log.d(RNVIEW_SHOT, "deleted file: " + file.getAbsolutePath());
+                    }
+                }
+            }
+        }
+    }
+
+    /**
+     * Create a temporary file in the cache directory on either internal or external storage,
+     * whichever is available and has more free space.
      */
     @NonNull
     private File createTempFile(@NonNull final Context context, @NonNull final String ext) throws IOException {
-        final File cacheDir = context.getCacheDir();
+        final File externalCacheDir = context.getExternalCacheDir();
+        final File internalCacheDir = context.getCacheDir();
+        final File cacheDir;
+
+        if (externalCacheDir == null && internalCacheDir == null) {
+            throw new IOException("No cache directory available");
+        }
+
+        if (externalCacheDir == null) {
+            cacheDir = internalCacheDir;
+        } else if (internalCacheDir == null) {
+            cacheDir = externalCacheDir;
+        } else {
+            cacheDir = externalCacheDir.getFreeSpace() > internalCacheDir.getFreeSpace() ?
+                    externalCacheDir : internalCacheDir;
+        }
 
         final String suffix = "." + ext;
         return File.createTempFile(TEMP_FILE_PREFIX, suffix, cacheDir);
